requirements.txt
fastapi==0.109.0
uvicorn==0.25.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
apscheduler==3.10.4
requests==2.31.0
pydantic==2.5.0
email-validator==2.1.0
